Don't encode url in the query parameters operation.
For example, if we have 'field[name]=a' in the query parameters
in the url it will be depicted as 'field%5Bname%5D=a' in the url textfield.
To make the url in the textfield as closer as in the query parameters
grid we don't need to encode the whole url.
Actual url encoding is happened just before a request sending.